requirements.txt
langchain_openai
langchain
langchain_core
python-dotenv
streamlit
langchain_community